+++ libs/libcommon/src/libcommon/utils.py
@@ -93,16 +93,16 @@ def get_datetime(days: Optional[float] = None) -> datetime:
     return date
 
 
-def get_duration(started_at: datetime) -> int:
+def get_duration(started_at: datetime) -> float:
     """
     Get time in seconds that has passed from `started_at` until now.
     `started_at` must be in UTC timezone.
     """
     started_at = pytz.UTC.localize(started_at) if not started_at.tzinfo else started_at
-    return int((get_datetime() - started_at).total_seconds())
+    return (get_datetime() - started_at).total_seconds()
 
 
-def get_duration_or_none(started_at: Optional[datetime]) -> Optional[int]:
+def get_duration_or_none(started_at: Optional[datetime]) -> Optional[float]:
     return get_duration(started_at) if started_at else None
